Port forwarding
Adds a local daemon to handle port requests.

The container is started with 1 or more open ports. The client can
forward these to their machine by running "rx ws open-port 2468". This
will start a daemon (if not already running) that can access the
workspace's localhost:2468 through the local machine's 2468 port.